Fix time-series feature math, harden data quality, clean up lint and hygiene
Metrics engine:
- computeDataQuality now judges reliability against the requested Window
  (new metrics.Window type) so a metric that stops reporting mid-window is
  flagged unreliable instead of looking complete; Process and
  ProcessWithBaselineStats take Window explicitly and every caller passes
  real bounds
- floor near-zero denominators in computeStepChange and computeTrend so a
  tiny non-zero baseline no longer explodes changePct or zeroes TrendScore
- extract classifyCore and saturation magic numbers into named constants
- document the two Stddev scales in ComputeRobustBaselineStats

Bug fixes:
- drop leaked d.Registry identifier from the metrics_list tool description
  and the metrics_snapshot baseline note
- assert the exact tool set (incl. errors_trends, trace_find_from_logs) in
  TestListTools instead of a presence-only check
- use safeInt32 for profiler and traces page sizes

Hygiene and lint:
- remove machine-specific rapid *.fail seeds and gitignore them
- pin alpine, run as non-root, add .dockerignore; add test targets to Makefile
- drive production code to zero golangci-lint issues (errorlint %w, US-locale
  spelling, ReadHeaderTimeout, wrapped yaml errors) and move conventional
  test-only noise into golangci exclusions

+++ internal/tools/metrics_chart_test.go
@@ -74,7 +74,7 @@ func TestToChartPoints(t *testing.T) {
 		}
 		got := toChartPoints(pts)
 		require.Len(t, got, 1)
-		assert.Equal(t, int64(now.Add(3*time.Minute).Unix()), got[0].TS)
+		assert.Equal(t, now.Add(3*time.Minute).Unix(), got[0].TS)
 		assert.Equal(t, 42.0, got[0].V)
 	})
